Statically link MesenCore
This means no copying `MesenCore.so` to a user directory, which is both
less chance for breakage (no potentially decoupling the `.so` from the
executable and getting ABI breakage), somewhat faster startup times
(no extracting that file or even checking its info), *and* possibly better
performance (since everything gets compiled together).

Speaking of: I tested, and performance of the resulting binary did not
regress; possibly improved, but negligibly so.